music/requirements.txt
pyrogram==2.0.106
pytgcalls==3.0.0.dev24
yt-dlp==2023.12.30
youtube-search-python==1.6.6
aiohttp==3.9.1
aiofiles==23.2.1
python-dotenv==1.0.0
psutil==5.9.6
requests==2.31.0
pillow==10.1.0
motor==3.3.2
dnspython==2.4.2
pytz==2023.3